requirements.txt
# -----------------------------------------
# 🔹 Core Deep Learning (PyTorch)
# -----------------------------------------
torch==2.2.0+cpu
torchvision==0.17.0+cpu
torchaudio==2.2.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

numpy
pandas
scikit-learn
tqdm

# -----------------------------------------
# 🔹 Image Processing
# -----------------------------------------
Pillow
opencv-python
rembg
colorthief
webcolors
imagehash

# -----------------------------------------
# 🔹 AI Image Generation (SDXL / FLUX)
# -----------------------------------------
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
safetensors==0.4.2
huggingface-hub==0.21.4

# -----------------------------------------
# 🔹 Web Application (UI)
# -----------------------------------------
streamlit==1.31.1
streamlit-image-select==0.6.0
python-dotenv

# -----------------------------------------
# 🔹 Visualization (Training Metrics)
# -----------------------------------------
matplotlib
seaborn

# -----------------------------------------
# 🔹 Utilities
# -----------------------------------------
requests
pathlib
